Do not force subnormal values to 0 in dec2x.
The dec2x function called stold, which internally calls strtold and then throws an exception if strtold sets errno. However, strtold may set errno to ERANGE for underflow but still return a meaningful non-zero value (generally a subnormal). The behavior of stold prevents us from getting the actual value in that case, and it is effectively forced to 0.

To avoid this problem, the code now just calls strtold directly. It checks that the whole string was processed, but otherwise makes no adjustments to the return value, even in overflow/underflow cases. The values returned from strtold should be suitable as-is, at least on reasonable platforms with IEEE floating point.

An example that was affected by this is the handling of "1.6810515715560467531E-4932" on an Intel Mac.